created all classes and their constructors

+++ app/src/main/java/item/AdventuringGear.java
@@ -3,7 +3,7 @@ package item;
 public class AdventuringGear extends Item {
     private String description;
 
-    public AdventuringGear(String name, int weight, Currency cost, String description) {
+    public AdventuringGear(String name, Double weight, Currency cost, String description) {
         super(name, weight, cost);
         this.description = description;
     }
