+++ app/src/main/java/item/Ammunition.java
@@ -3,7 +3,7 @@ package item;
 public class Ammunition extends Item {
     private int amount;
 
-    public Ammunition(String name, int weight, Currency cost, int amount) {
+    public Ammunition(String name, Double weight, Currency cost, int amount) {
         super(name, weight, cost);
         this.amount = amount;
     }
